Filter out non cacheable requests from logs
